# Minor performance optimizations in Loader#cperf